pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiagent-debate"
version = "1.0.0"
description = "Production-grade multi-agent debate engine (Python CLI, OpenAI backend)"
authors = [{name = "You"}]
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1"
]

[project.scripts]
ma-debate = "ma_debate.cli:app"
